Remove the AugmentedThrowable type
